+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - mod：优化链接
 - random：增加泛型
+- decimal：增加 `Multiply` 方法
 
 ## v1.0.39 / 2021-12-27